Refactoring ingest code
Added separate payload/metadata options for ingesting data. Will be
useful when other datatypes will be ingested in the future (control
points, etc...).

+++ voxel_globe/ingest/views.py
@@ -4,6 +4,9 @@ from django.shortcuts import render_to_response
 from django.template import RequestContext
 
 from .forms import UploadFileForm
+from .tools import METADATA_TYPES, PAYLOAD_TYPES
+#Deprecating ...
+from .tools import SENSOR_TYPES
 
 from voxel_globe.ingest import models
 
@@ -30,41 +33,16 @@ def ViewSetFactory(model, serializer):
   return type('ViewSet_%s' % model._meta.model_name, (IngestViewSet,), {'queryset':model.objects.all(), 'serializer_class':serializer})
 
 router.register(models.File._meta.model_name, ViewSetFactory(models.File, voxel_globe.ingest.serializers.FileSerializer))
-router.register(models.Directory._meta.model_name, ViewSetFactory(models.Directory, voxel_globe.ingest.serializers.DirectorySerializer))
-router.register(models.Directory._meta.model_name+'_nest', ViewSetFactory(models.Directory, voxel_globe.ingest.serializers.NestFactory(voxel_globe.ingest.serializers.DirectorySerializer)))
+#router.register(models.Directory._meta.model_name, ViewSetFactory(models.Directory, voxel_globe.ingest.serializers.DirectorySerializer))
+#router.register(models.Directory._meta.model_name+'_nest', ViewSetFactory(models.Directory, voxel_globe.ingest.serializers.NestFactory(voxel_globe.ingest.serializers.DirectorySerializer)))
 router.register(models.UploadSession._meta.model_name, ViewSetFactory(models.UploadSession, voxel_globe.ingest.serializers.UploadSessionSerializer));
-router.register(models.UploadSession._meta.model_name+'_nest', ViewSetFactory(models.UploadSession, voxel_globe.ingest.serializers.NestFactory(voxel_globe.ingest.serializers.UploadSessionSerializer)));
-
-#key: [Friendly name, moduleName]
-#Module name should not include tasks, but it is assume that tasks.ingest_data is used
-#I'm sure this will be updated at a later time to have api data in the module rather than here 
-#SENSOR_TYPES = {'arducopter':'Arducopter', 
-#                'jpg_exif':'JPEG with EXIF tags'};
-#to be used in conjunction with importlib
-
-def getSensorTypes():
-  ''' Helper function to get all registered ingest functions '''
-  class IngestClass(object):
-    def __init__(self, ingest_data, description=''):
-      self.ingest_data=ingest_data
-      self.description=description
-  ingests = {}
-  import django.conf
-  import importlib
-  for tasks in django.conf.settings.INSTALLED_APPS:
-    try:
-      mod = importlib.import_module(tasks+'.tasks')
-      task = mod.ingest_data
-      ingests[task.dbname] = IngestClass(task, task.description)
-    except (ImportError, AttributeError):
-      pass
-  return ingests
-
-SENSOR_TYPES = getSensorTypes()
+#router.register(models.UploadSession._meta.model_name+'_nest', ViewSetFactory(models.UploadSession, voxel_globe.ingest.serializers.NestFactory(voxel_globe.ingest.serializers.UploadSessionSerializer)));
 
 def chooseSession(request):
     return render_to_response('ingest/html/chooseSession.html', 
-                            {'sensorTypes': SENSOR_TYPES}, 
+                            {'sensorTypes': SENSOR_TYPES,
+                             'payload_types': PAYLOAD_TYPES,
+                             'metadata_types': METADATA_TYPES}, 
                             context_instance=RequestContext(request))
 
 def addFiles(request):
@@ -72,45 +50,45 @@ def addFiles(request):
   uploadSession = models.UploadSession.objects.get(id=upload_session_id)
   
   #Temporary code only. This will get one directory successfully, or create a new one.
-  try:
-    directory = uploadSession.directory.get()
-  except:
-    directory = models.Directory(name='Blahdir', session=uploadSession, owner=request.user);
-    directory.save();
-    directory.name = str(directory.id); directory.save();
-  testFile = models.File(name='Blahfile', directory=directory, owner=request.user);
+###  try:
+###    directory = uploadSession.directory.get()
+###  except:
+###    directory = models.Directory(name='Blahdir', session=uploadSession, owner=request.user);
+###    directory.save();
+###    directory.name = str(directory.id); directory.save();
+  testFile = models.File(name='Newfile', session=uploadSession, owner=request.user);
   testFile.save();
 
   return render_to_response('ingest/html/addFiles.html',
                            {'uploadSession':uploadSession,
-                            'directory':directory,
+###                            'directory':directory,
                             'testFile':testFile}, 
                             context_instance=RequestContext(request))
 
-def blah(request):
-  uploadSession = models.UploadSession(name='Blah', owner=request.user);
-  uploadSession.save();
-  uploadSession.name = str(uploadSession.id); uploadSession.save();
-  directory = models.Directory(name='Blahdir', session=uploadSession, owner=request.user);
-  directory.save();
-  directory.name = str(directory.id); directory.save();
-  testFile = models.File(name='Blahfile', directory=directory, owner=request.user);
-  testFile.save();
+# def blah(request):
+#   uploadSession = models.UploadSession(name='Blah', owner=request.user);
+#   uploadSession.save();
+#   uploadSession.name = str(uploadSession.id); uploadSession.save();
+#   directory = models.Directory(name='Blahdir', session=uploadSession, owner=request.user);
+#   directory.save();
+#   directory.name = str(directory.id); directory.save();
+#   testFile = models.File(name='Blahfile', directory=directory, owner=request.user);
+#   testFile.save();
   
-  if request.method=='POST':
-    form = UploadFileForm(request.POST, request.FILES);
-    if form.is_valid():
-      return HttpResponse('form valid');
-  else:
-    form = UploadFileForm();
+#   if request.method=='POST':
+#     form = UploadFileForm(request.POST, request.FILES);
+#     if form.is_valid():
+#       return HttpResponse('form valid');
+#   else:
+#     form = UploadFileForm();
 
-  #return render(request, 'ingest/html/upload.html', {'form':form})
-  return render_to_response('ingest/html/upload.html', 
-                            {'form':form,
-                             'uploadSession':uploadSession,
-                             'directory':directory,
-                             'testFile':testFile}, 
-                            context_instance=RequestContext(request))
+#   #return render(request, 'ingest/html/upload.html', {'form':form})
+#   return render_to_response('ingest/html/upload.html', 
+#                             {'form':form,
+#                              'uploadSession':uploadSession,
+#                              'directory':directory,
+#                              'testFile':testFile}, 
+#                             context_instance=RequestContext(request))
   
 def upload(request):
   try:
@@ -120,10 +98,10 @@ def upload(request):
     uploadSession.save();
     uploadSession.name = str(uploadSession.id); uploadSession.save();
     uploadSession_id = uploadSession.id
-  try:
-    directory_id = request.POST['directory']
-  except:
-    directory_id = 'failsafe'
+#  try:
+#    directory_id = request.POST['directory']
+#  except:
+#    directory_id = 'failsafe'
   try:
     testFile_id = request.POST['testFile']
   except:
@@ -131,7 +109,7 @@ def upload(request):
 
   s = 'ok<br>'
   
-  saveDir = os.path.join(os.environ['VIP_TEMP_DIR'], str(uploadSession_id), str(directory_id))
+  saveDir = os.path.join(os.environ['VIP_TEMP_DIR'], 'ingest', str(uploadSession_id))
   distutils.dir_util.mkpath(saveDir)
   
   for f in request.FILES:
@@ -143,28 +121,33 @@ def upload(request):
   return HttpResponse(s);
 
 def ingestFolder(request):
+  from vsi.tools.dir_util import mkdtemp
+  from celery.canvas import chain
+
   uploadSession_id = request.POST['uploadSession']
   #directories = models.Directory.objects.filter(uploadSession_id = uploadSession_id)
   #Code not quite done, using failsafe for now. 
   uploadSession = models.UploadSession.objects.get(id=uploadSession_id);
 
-  sessionDir = os.path.join(os.environ['VIP_TEMP_DIR'], str(uploadSession.id))
-  imageDir = os.path.join(os.environ['VIP_IMAGE_SERVER_ROOT'], str(uploadSession.id))
-  if os.path.exists(imageDir):
-    from vsi.tools.dir_util import mkdtemp
-    imageDir = mkdtemp(dir=os.environ['VIP_IMAGE_SERVER_ROOT']);
+  sessionDir = os.path.join(os.environ['VIP_TEMP_DIR'], 'ingest', str(uploadSession.id))
+  #imageDir = os.path.join(os.environ['VIP_IMAGE_SERVER_ROOT'], str(uploadSession.id))
+  #if os.path.exists(imageDir):
+  imageDir = mkdtemp(dir=os.environ['VIP_IMAGE_SERVER_ROOT'], prefix='img');
   
-  print sessionDir, imageDir
-  #TODO: The following should be a celery task too, and the two should be a 
-  #workflow, and this plus ingest_data should be a workflow
-
-  from vsi.iglob import glob  
-  metadata = glob(sessionDir+'/*/*_adj_tagged_images.txt', False);
   
-  distutils.dir_util.copy_tree(sessionDir, imageDir)
-  distutils.dir_util.remove_tree(sessionDir)
-
-  task = SENSOR_TYPES[uploadSession.sensorType].ingest_data.delay(uploadSession_id, imageDir);
+  #Deprecated code
+  if uploadSession.sensorType != "None":
+    distutils.dir_util.copy_tree(sessionDir, imageDir)
+    distutils.dir_util.remove_tree(sessionDir)
+    task = SENSOR_TYPES[uploadSession.sensorType].ingest_data.delay(uploadSession_id, imageDir)
+  else:
+    #This is the REAL non-deprecated code. Not done yet ;)
+    import voxel_globe.ingest.tasks
+    task0 = voxel_globe.ingest.tasks.move_data.si(sessionDir, imageDir)
+    task1 = PAYLOAD_TYPES[uploadSession.payload_type].ingest.si(uploadSession_id, imageDir)
+    task2 = METADATA_TYPES[uploadSession.metadata_type].ingest.si(uploadSession_id, imageDir)
+    tasks = task0 | task1 | task2 #create chain
+    tasks.apply_async()
 
   return render(request, 'ingest/html/ingest_started.html', 
                 {'task_id':task.task_id})
